Python_Code/requirements.txt
psycopg2-binary
folium
requests
